pyproject.toml
[project]
name = "rudebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.15",
    "fastapi>=0.116.1",
    "openai>=1.100.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "uvicorn>=0.35.0",
]
